Move route create/edit/delete functionality to own component
