+++ src/components/map/ModalMap.tsx
@@ -1,39 +1,16 @@
-import React, { useContext, useEffect, useRef, useState } from 'react';
-import { useTranslation } from 'react-i18next';
+import React, { useEffect, useRef } from 'react';
 import { useHistory } from 'react-router';
-import { MapFilterContext, setObservationDate } from '../../context/MapFilter';
-import { RouteRoute } from '../../generated/graphql';
+import { useAppDispatch, useAppSelector, useMapUrlQuery } from '../../hooks';
 import {
-  useAppDispatch,
-  useAppSelector,
-  useDeleteRoute,
-  useEditRouteGeometry,
-  useExtractRouteFromFeature,
-  useMapUrlQuery,
-} from '../../hooks';
-import { useCreateRoute } from '../../hooks/routes/useCreateRoute';
-import {
-  initializeMapEditorWithRoutesAction,
   resetMapEditorStateAction,
   selectIsModalMapOpen,
-  selectMapEditor,
   setIsModalMapOpenAction,
-  setSelectedRouteIdAction,
-  stopDrawRouteAction,
-  toggleDrawRouteAction,
-  toggleEditRouteAction,
 } from '../../redux';
-import { isDateInRange } from '../../time';
-import { RequiredKeys } from '../../types';
-import { ConfirmationDialog, Modal } from '../../uiComponents';
-import { showSuccessToast } from '../../utils';
-import {
-  ConflictResolverModal,
-  mapRouteToCommonConflictItem,
-} from '../routes-and-lines/common/ConflictResolverModal';
+import { Modal } from '../../uiComponents';
 import { Map } from './Map';
 import { MapFooter } from './MapFooter';
 import { MapHeader } from './MapHeader';
+import { RouteEditor } from './routes/RouteEditor';
 
 interface Props {
   className?: string;
@@ -47,47 +24,12 @@ const mapFooterHeight = 82;
 
 export const ModalMap: React.FC<Props> = ({ className }) => {
   const mapRef = useRef<ExplicitAny>(null);
-  const { t } = useTranslation();
+  const routeEditorRef = useRef<ExplicitAny>(null);
+
   const dispatch = useAppDispatch();
   const history = useHistory();
   const isModalMapOpen = useAppSelector(selectIsModalMapOpen);
   const { deleteMapOpenQueryParameter, isMapOpen } = useMapUrlQuery();
-
-  const { editedRouteData, creatingNewRoute, initiallyDisplayedRouteIds } =
-    useAppSelector(selectMapEditor);
-
-  const {
-    state: { observationDate },
-    dispatch: mapFilterDispatch,
-  } = useContext(MapFilterContext);
-
-  const {
-    id: editedRouteId,
-    infraLinks,
-    stops: routeStops,
-    metaData: routeDetails,
-  } = editedRouteData;
-
-  const [isDeleting, setIsDeleting] = useState(false);
-
-  const {
-    prepareCreate,
-    mapCreateChangesToVariables,
-    insertRouteMutation,
-    defaultErrorHandler: defaultInsertRouteErrorHandler,
-  } = useCreateRoute();
-
-  const {
-    prepareEditGeometry,
-    mapEditGeometryChangesToVariables,
-    editRouteGeometryMutation,
-  } = useEditRouteGeometry();
-
-  const [conflicts, setConflicts] = useState<RouteRoute[]>([]);
-  const { deleteRoute, defaultErrorHandler: defaultDeleteErrorHandler } =
-    useDeleteRoute();
-
-  const { mapRouteStopsToStopIds } = useExtractRouteFromFeature();
 
   const syncIsModalMapStateWithMapOpenQueryParam = () => {
     const mapOpen = isMapOpen();
@@ -108,138 +50,6 @@ export const ModalMap: React.FC<Props> = ({ className }) => {
       syncIsModalMapStateWithMapOpenQueryParam();
     });
   });
-
-  const editRoute = async (routeId: UUID) => {
-    const stopIdsWithinRoute = mapRouteStopsToStopIds(routeStops);
-
-    const changes = await prepareEditGeometry({
-      routeId,
-      newGeometry: {
-        stopIdsWithinRoute,
-        infraLinksAlongRoute: infraLinks || [],
-      },
-    });
-
-    const variables = mapEditGeometryChangesToVariables(changes);
-
-    await editRouteGeometryMutation(variables);
-  };
-
-  const createRoute = async () => {
-    const stopIdsWithinRoute = mapRouteStopsToStopIds(routeStops);
-
-    const changes = await prepareCreate({
-      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-      form: routeDetails!,
-      routeGeometry: {
-        stopIdsWithinRoute,
-        infraLinksAlongRoute: infraLinks || [],
-      },
-    });
-    if (changes.conflicts?.length) {
-      setConflicts(changes.conflicts);
-      return undefined;
-    }
-    const variables = mapCreateChangesToVariables(changes);
-
-    const response = await insertRouteMutation(variables);
-
-    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-    return response.data!.insert_route_route_one!;
-  };
-
-  const onRouteCreated = (
-    newRoute: RequiredKeys<Partial<RouteRoute>, 'route_id'>,
-  ) => {
-    // Add created route to displayed routes
-    dispatch(
-      initializeMapEditorWithRoutesAction([
-        ...(initiallyDisplayedRouteIds || []),
-        newRoute.route_id,
-      ]),
-    );
-
-    // Select created route
-    dispatch(setSelectedRouteIdAction(newRoute.route_id));
-
-    // If created route is not valid at the selected observation date,
-    // change observation date to created route's validity start date
-    // so the user can see the freshly created route
-
-    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-    const validityStart = newRoute.validity_start!;
-
-    if (
-      !isDateInRange(observationDate, validityStart, newRoute?.validity_end)
-    ) {
-      mapFilterDispatch(setObservationDate(validityStart));
-      showSuccessToast(t('filters.observationDateAdjusted'));
-    }
-
-    // Reset map editor drap mode and remove draft route
-    // as it is now saved
-
-    dispatch(stopDrawRouteAction());
-
-    mapRef?.current?.onDeleteDrawnRoute();
-  };
-
-  const onDrawRoute = () => {
-    mapRef?.current?.onDeleteDrawnRoute();
-    dispatch(toggleDrawRouteAction());
-  };
-  const onEditRoute = () => {
-    if (!creatingNewRoute) {
-      mapRef?.current?.onDeleteDrawnRoute();
-    }
-    dispatch(toggleEditRouteAction());
-  };
-  const onCancel = () => {
-    mapRef?.current?.onDeleteDrawnRoute();
-
-    dispatch(stopDrawRouteAction());
-  };
-  const onSave = async () => {
-    try {
-      if (editedRouteId) {
-        await editRoute(editedRouteId);
-        showSuccessToast(t('routes.saveSuccess'));
-      } else {
-        const createdRoute = await createRoute();
-        showSuccessToast(t('routes.saveSuccess'));
-
-        if (createdRoute) {
-          onRouteCreated(createdRoute);
-        }
-      }
-    } catch (err) {
-      defaultInsertRouteErrorHandler(err);
-    }
-  };
-  const onDeleteConfirm = async () => {
-    try {
-      if (!editedRouteId) {
-        return;
-      }
-
-      // delete the route from the backend
-      await deleteRoute(editedRouteId);
-      showSuccessToast(t('routes.deleteSuccess'));
-
-      setIsDeleting(false);
-      dispatch(setIsModalMapOpenAction(false));
-    } catch (err) {
-      defaultDeleteErrorHandler(err);
-    }
-  };
-
-  const onDeleteRoute = async () => {
-    if (creatingNewRoute) {
-      onCancel();
-    } else {
-      setIsDeleting(true);
-    }
-  };
 
   const onCloseModalMap = () => {
     dispatch(resetMapEditorStateAction());
@@ -269,24 +79,15 @@ export const ModalMap: React.FC<Props> = ({ className }) => {
         ref={mapRef}
       />
       <MapFooter
-        onDrawRoute={onDrawRoute}
-        onEditRoute={onEditRoute}
-        onDeleteRoute={onDeleteRoute}
-        onCancel={onCancel}
-        onSave={onSave}
+        onDrawRoute={() => routeEditorRef.current.onDrawRoute()}
+        onEditRoute={() => routeEditorRef.current.onEditRoute()}
+        onDeleteRoute={() => routeEditorRef.current.onDeleteRoute()}
+        onCancel={() => routeEditorRef.current.onCancel()}
+        onSave={() => routeEditorRef.current.onSave()}
       />
-      <ConfirmationDialog
-        isOpen={isDeleting}
-        onCancel={() => setIsDeleting(false)}
-        onConfirm={onDeleteConfirm}
-        title={t('confirmDeleteRouteDialog.title')}
-        description={t('confirmDeleteRouteDialog.description')}
-        confirmText={t('confirmDeleteRouteDialog.confirmText')}
-        cancelText={t('cancel')}
-      />
-      <ConflictResolverModal
-        onClose={() => setConflicts([])}
-        conflicts={conflicts.map(mapRouteToCommonConflictItem)}
+      <RouteEditor
+        onDeleteDrawnRoute={() => mapRef.current.onDeleteDrawnRoute()}
+        ref={routeEditorRef}
       />
     </Modal>
   );
